refactor: remove shallow tracer struct
This commit removes the `AccessListTracer` which was shallow. Now our EVM tracer directly uses Alloy.
